production

+++ delete.sh
@@ -4,8 +4,11 @@
 num_iterations=1
 
 for ((i = 1; i <= num_iterations; i++)); do
-    # Run the command
-    rm backend_v3/.env
+
+    # Check if the file exists before removing it
+    if [ -e backend_v3/.env ]; then
+        rm backend_v3/.env
+    fi
     git add .
     git commit --amend --no-edit
     git rebase --continue
